Added in new data set
Testing a better setup for base stats. These can be accessed inside each hero data set using the key: "stats2"

This set of data doesn't specify the name of the stat as a key, and instead uses an array of objects that declare a name, along with other details about the stat. The idea here was to be able to include crystal damage, special circumstances stats (like bonus stats), extra stats from perks, etc. all into the same schema as each heroes base stats. This requires some additional data for CP and WP scaling, as well as some basic type declarations that simply classify the type of base stat so they aren't completely abstracted when being used and can be grouped for various reasons.

Note: The values for a base hero stat "type" are currently intended to be:
"Base", "Crystal", "Weapon", and "Bonus"