MAJOR REVISION: HEROES DATA STRUCTURE
Added in the new data structure for heroes. This new json object (with the current key of "heroes2") will replace the first iteration "heroes" object, which is still in the file for performance testing. This will be removed within the next few days, so if using the data currently, so please be aware of this future breaking change. 

Otherwise, please use the "heroes2" data set and be ready to change the string "heroes2" to just "heroes" following the change.